fix: add ::bigint cast for gate_session_clips.session_id comparisons
gate_session_clips.session_id is BIGINT, but gate_sessions.door_id is TEXT
(the view casts session_id::text). Comparing them without a cast causes
"operator does not exist: bigint = text" — both the backfill and the
worker poll were silently skipping all outgoing sessions, so no clips
were ever fetched and face recognition could never run.

Affected: LEFT JOIN LATERAL and EXISTS checks in both the API backfill
endpoint and worker poll_new_gate_events.

+++ services/api/app/routers/enroll.py
@@ -401,7 +401,7 @@ def backfill(body: dict, _=Depends(require_auth)):
                     LEFT JOIN LATERAL (
                         SELECT gsc.label
                         FROM gate_session_clips gsc
-                        WHERE gsc.session_id = gs.door_id
+                        WHERE gsc.session_id = gs.door_id::bigint
                           AND gsc.label LIKE 'P.%%'
                           {room_join_filter}
                         LIMIT 1
@@ -411,7 +411,7 @@ def backfill(body: dict, _=Depends(require_auth)):
                       -- Phải có ít nhất 1 clip (dù label rỗng)
                       AND EXISTS (
                           SELECT 1 FROM gate_session_clips gsc2
-                          WHERE gsc2.session_id = gs.door_id
+                          WHERE gsc2.session_id = gs.door_id::bigint
                             AND gsc2.direction  = 'outgoing'
                       )
                       -- Bỏ qua nếu đang pending/running
